restructure files for clearer usability
