+++ prompt_generator/utils/__init__.py
@@ -1,0 +1,1 @@
+# This is the utils package file
